default offence state improvements
